Move and rename iterator class. Add strict types.
